requirements.txt
aiohttp>=3.9.0
pydantic>=2.5.0
pyyaml>=6.0.1
pytest>=7.4.0
pytest-asyncio>=0.23.0
aioresponses>=0.7.6
